raibis/pyproject.toml
[project]
name = "raibis"
version = "0.1.0"
description = "Personal LifeOS — TUI + AI, open source"
requires-python = ">=3.11"
dependencies = [
    "textual>=0.61.0",
    "anthropic>=0.25.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",        # stub — install when needed
    "httpx>=0.27.0",        # used by ollama provider
]

[project.scripts]
raibis = "raibis.tui.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["raibis"]
